Clean up and review feedback
