+++ tools/compat/errno.rs
@@ -15,7 +15,9 @@ cfg_if::cfg_if! {
             libc::__errno()
         }
     } else if #[cfg(target_os = "windows")] {
-        extern "C" { pub fn _errno() -> *mut c_int; }
+        extern "C" {
+            pub fn _errno() -> *mut c_int;
+        }
 
         pub unsafe fn errno_location() -> *mut c_int {
             _errno()
