+++ tools/dav1d_cli_parse.rs
@@ -1,16 +1,8 @@
-#[cfg(target_os = "windows")]
 use crate::compat::getopt::getopt_long;
-#[cfg(target_os = "windows")]
 use crate::compat::getopt::option;
-#[cfg(target_os = "windows")]
-use crate::compat::stdio::sprintf;
 use crate::compat::stdio::stderr;
 use libc::fprintf;
-#[cfg(not(target_os = "windows"))]
-use libc::getopt_long;
 use libc::memset;
-#[cfg(not(target_os = "windows"))]
-use libc::option;
 use libc::strcat;
 use libc::strcmp;
 use libc::strcpy;
